Refactoring - исправила замечания ревьюера
