+++ FakeNFT/Resources/Assets.xcassets/56_digital_art_x4.imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "56_digital_art_x4@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
