+++ FakeNFT/Resources/Assets.xcassets/ApeCoin (APE).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "ApeCoin (APE)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
