+++ FakeNFT/Resources/Assets.xcassets/Bitcoin (BTC).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Bitcoin (BTC).svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
