+++ FakeNFT/Resources/Assets.xcassets/Cardano (ADA).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Cardano (ADA)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
