+++ FakeNFT/Resources/Assets.xcassets/Dogecoin (DOGE).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Dogecoin (DOGE)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
