+++ FakeNFT/Resources/Assets.xcassets/Ethereum (ETH).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Ethereum (ETH)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
