+++ FakeNFT/Resources/Assets.xcassets/Shiba Inu (SHIB).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Shiba Inu (SHIB)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
