+++ FakeNFT/Resources/Assets.xcassets/Solana (SOL).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Solana (SOL)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
