+++ FakeNFT/Resources/Assets.xcassets/Tether (USDT).imageset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "Tether (USDT)@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
