+++ FakeNFT/Scenes /Cart/CartViewModel.swift
@@ -1,13 +1,5 @@
 import UIKit
 
-enum MockData {
-    static let items: [NFTItem] = [
-        NFTItem(image: UIImage(named: "nft1"), title: "April", rating: 1, price: 1.78),
-        NFTItem(image: UIImage(named: "nft2"), title: "Greena", rating: 3, price: 1.79),
-        NFTItem(image: UIImage(named: "nft3"), title: "Spring", rating: 5, price: 1.77)
-    ]
-}
-
 final class CartViewModel {
     
     // MARK: - Output bindings
