+++ FakeNFT/Scenes /Cart/MockData.swift
@@ -1,0 +1,22 @@
+import UIKit
+
+enum MockData {
+    static let items: [NFTItem] = [
+        NFTItem(image: UIImage(named: "nft1"), title: "April", rating: 1, price: 1.78),
+        NFTItem(image: UIImage(named: "nft2"), title: "Greena", rating: 3, price: 1.79),
+        NFTItem(image: UIImage(named: "nft3"), title: "Spring", rating: 5, price: 1.77)
+    ]
+}
+
+enum MockCrypto {
+    static let items: [CryptoPayment] = [
+        .init(iconName: "Bitcoin (BTC)", title: "Bitcoin", ticker: "BTC"),
+        .init(iconName: "Dogecoin (DOGE)", title: "Dogecoin", ticker: "DOGE"),
+        .init(iconName: "Tether (USDT)", title: "Tether", ticker: "USDT"),
+        .init(iconName: "ApeCoin (APE)", title: "ApeCoin", ticker: "APE"),
+        .init(iconName: "Solana (SOL)", title: "Solana", ticker: "SOL"),
+        .init(iconName: "Ethereum (ETH)", title: "Ethereum", ticker: "ETH"),
+        .init(iconName: "Cardano (ADA)", title: "Cardano", ticker: "ADA"),
+        .init(iconName: "Shiba Inu (SHIB)", title: "Shiba Inu", ticker: "SHIB")
+    ]
+}
